add map return

+++ new_user/views.py
@@ -18,7 +18,7 @@ week_change = {'1': '一', '2': '二', '3': '三', '4': '四', '5': '五', '6': 
 #     name = forms.CharField(min_length=4, label='用户名', required=True, help_text='必填')   # 必须用required
 #     pwd = forms.CharField(min_length=4, label='密码', required=True, help_text='必填')
 #     weChat = forms.CharField(min_length=4, label='微信号', required=True, help_text='必填')
-#
+
 
 # 用户登录
 def loginView(request):    # 设置标题和另外两个URL链接
@@ -113,6 +113,7 @@ def activityView(request, number):
     activity_time = str(activity_start_time) + '-' + str(activity_end_time) + ' 周' + week_change[str(activity_week)]
     activity_place = str(BadmintonActivity.objects.get(activity_number=int(number)).activity_place)
     join_count = str(BadmintonActivityDetails.objects.filter(activity_number_id=int(number)).count())
+    is_operate = BadmintonActivity.objects.get(activity_number=int(number)).is_operate
     surplus = str(12 - int(join_count))
     logger.info(activity_place)
     for i in activityDetails:
@@ -127,9 +128,12 @@ def activityView(request, number):
             join_person = BadmintonActivityDetails(join_weChat_id=request.user.id, activity_number_id=int(number))
             join_person.save()
         elif action == 'cancel':
-            tips = '取消报名成功'
-            cancel_activity = BadmintonActivityDetails.objects.filter(activity_number_id=int(number), join_weChat_id=request.user.id).delete()
-            # logger.info(txt1)
+            if is_operate:
+                tips = '取消报名成功！'
+                cancel_activity = BadmintonActivityDetails.objects.filter(activity_number_id=int(number), join_weChat_id=request.user.id).delete()
+                # logger.info(txt1)
+            else:
+                tips = '活动开始前一天不允许取消报名！！！'
         elif action == 'substitution':
             tips = '替补成功'
             substitution_activity = BadmintonActivityDetails(join_weChat_id=request.user.id, activity_number_id=int(number),
